#152: transition counter 성능 최적화

+++ src/main/kotlin/kr/co/plasticcity/kotlmata/keywords.kt
@@ -117,7 +117,7 @@ interface PayloadHolder
 
 interface TransitionHolder
 {
-	val count: Int
+	val count: Long
 }
 
 interface ErrorActionDSL : ErrorHolder, ActionDSL
